libft_m

+++ 42cursus/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: haekang <haekang@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 19:44:14 by haekang           #+#    #+#             */
-/*   Updated: 2023/03/29 05:25:07 by haekang          ###   ########.fr       */
+/*   Updated: 2023/04/13 17:11:21 by haekang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
